Move sortable-column component to root level
