+++ src/ggrc/assets/javascripts/components/sortable-column/sortable-column.js
@@ -3,7 +3,7 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import template from './templates/sortable-column.mustache';
+import template from './sortable-column.mustache';
 
 GGRC.Components('sortableColumn', {
   tag: 'sortable-column',
