+++ src/ggrc/assets/javascripts/components/sortable-column/sortable-column.mustache
